backend/requirements.txt
Flask==2.3.2
redis==4.5.5
celery==5.3.1
psycopg2-binary==2.9.6
